organize vault hooks
