+++ src/components/Vault/APR/CurveBalancerApr.tsx
@@ -8,7 +8,7 @@ import {
   useVaultExtraApr,
   useVaultPoolId,
   useVaultTotalApr,
-} from "@/hooks/data"
+} from "@/hooks/data/vaults"
 
 import Percentage from "@/components/Percentage"
 import Skeleton from "@/components/Skeleton"
