+++ src/components/Vault/VaultDepositForm.tsx
@@ -16,9 +16,8 @@ import { toFixed } from "@/lib/api/util/format"
 import isEthTokenAddress from "@/lib/isEthTokenAddress"
 import logger from "@/lib/logger"
 import { VaultProps } from "@/lib/types"
-import { useVaultPoolId } from "@/hooks/data"
 import { usePreviewDeposit } from "@/hooks/data/preview/usePreviewDeposit"
-import { useVault } from "@/hooks/data/vaults"
+import { useVault, useVaultPoolId } from "@/hooks/data/vaults"
 import useActiveChainId from "@/hooks/useActiveChainId"
 import useTokenOrNative from "@/hooks/useTokenOrNative"
 import { useIsTokenCompounder } from "@/hooks/useVaultTypes"
