+++ src/components/Vault/VaultWithdrawForm.tsx
@@ -13,9 +13,8 @@ import {
 import { toFixed } from "@/lib/api/util/format"
 import logger from "@/lib/logger"
 import { VaultProps } from "@/lib/types"
-import { useVaultPoolId } from "@/hooks/data"
 import { usePreviewRedeem } from "@/hooks/data/preview/usePreviewRedeem"
-import { useVault } from "@/hooks/data/vaults"
+import { useVault, useVaultPoolId } from "@/hooks/data/vaults"
 import useActiveChainId from "@/hooks/useActiveChainId"
 import useTokenOrNative from "@/hooks/useTokenOrNative"
 import { useIsTokenCompounder } from "@/hooks/useVaultTypes"
