--- src/hooks/data/index.ts
@@ -1,14 +1,0 @@
-export {
-  useVaultAuraApr,
-  useVaultBalApr,
-  useVaultBaseApr,
-  useVaultCrvApr,
-  useVaultCvxApr,
-  useVaultEthApr,
-  useVaultExtraApr,
-  useVaultGmxApr,
-  useVaultTotalApr,
-} from "./useVaultApy"
-export { default as useVaultDepositedAssets } from "./useVaultDepositedAssets"
-export { default as useVaultPoolId } from "./useVaultPoolId"
-export { default as useVaultTvl } from "./useVaultTvl"
