+++ src/hooks/data/vaults/fallbacks/useBalancerVaultGraphTotalApr.ts
