+++ src/hooks/data/vaults/fallbacks/useCurveVaultGraphTotalApr.ts
