+++ src/hooks/data/vaults/fallbacks/useTokenVaultGraphTotalApr.ts
@@ -1,8 +1,8 @@
 import { VaultDynamicProps } from "@/lib/types"
-import useTokenAuraBalVault from "@/hooks/data/aprFallbacks/useTokenAuraBalVault"
-import useTokenCvxCrvVault from "@/hooks/data/aprFallbacks/useTokenCvxCrvVault"
-import useTokenGlpVault from "@/hooks/data/aprFallbacks/useTokenGlpVault"
-import useTokenVaultSymbol from "@/hooks/data/aprFallbacks/useTokenVaultSymbol"
+import useTokenAuraBalVault from "@/hooks/data/vaults/fallbacks/useTokenAuraBalVault"
+import useTokenCvxCrvVault from "@/hooks/data/vaults/fallbacks/useTokenCvxCrvVault"
+import useTokenGlpVault from "@/hooks/data/vaults/fallbacks/useTokenGlpVault"
+import useTokenVaultSymbol from "@/hooks/data/vaults/fallbacks/useTokenVaultSymbol"
 
 export default function useTokenVaultGraphTotalApr({
   asset,
