+++ src/hooks/data/vaults/fallbacks/useTokenVaultSymbol.ts
