+++ src/hooks/data/vaults/index.ts
@@ -1,2 +1,5 @@
 export * from "./useVault"
+export * from "./useVaultApy"
 export * from "./useVaultFees"
+export * from "./useVaultPoolId"
+export * from "./useVaultTvl"
