+++ src/hooks/data/vaults/useVaultApy.ts
@@ -1,22 +1,18 @@
 import { VaultDynamicProps } from "@/lib/types"
 import { useApiVaultDynamic } from "@/hooks/api"
-import useBalancerVaultGraphTotalApr from "@/hooks/data/aprFallbacks/useBalancerVaultGraphTotalApr"
-import useCurveVaultGraphTotalApr from "@/hooks/data/aprFallbacks/useCurveVaultGraphTotalApr"
-import useTokenAuraBalVault from "@/hooks/data/aprFallbacks/useTokenAuraBalVault"
-import useTokenGlpVault from "@/hooks/data/aprFallbacks/useTokenGlpVault"
-import useTokenVaultGraphTotalApr from "@/hooks/data/aprFallbacks/useTokenVaultGraphTotalApr"
-import useTokenVaultSymbol from "@/hooks/data/aprFallbacks/useTokenVaultSymbol"
-import { useVaultAprFallback } from "@/hooks/data/aprFallbacks/useVaultAprFallback"
+import useBalancerVaultGraphTotalApr from "@/hooks/data/vaults/fallbacks/useBalancerVaultGraphTotalApr"
+import useCurveVaultGraphTotalApr from "@/hooks/data/vaults/fallbacks/useCurveVaultGraphTotalApr"
+import useTokenAuraBalVault from "@/hooks/data/vaults/fallbacks/useTokenAuraBalVault"
+import useTokenGlpVault from "@/hooks/data/vaults/fallbacks/useTokenGlpVault"
+import useTokenVaultGraphTotalApr from "@/hooks/data/vaults/fallbacks/useTokenVaultGraphTotalApr"
+import useTokenVaultSymbol from "@/hooks/data/vaults/fallbacks/useTokenVaultSymbol"
+import { useVaultAprFallback } from "@/hooks/data/vaults/fallbacks/useVaultAprFallback"
 import {
   useIsCurveCompounder,
   useIsTokenCompounder,
 } from "@/hooks/useVaultTypes"
 
-export default function useVaultApy({
-  asset,
-  poolId,
-  type,
-}: VaultDynamicProps) {
+export function useVaultApy({ asset, poolId, type }: VaultDynamicProps) {
   const isCurve = useIsCurveCompounder(type)
   const isToken = useIsTokenCompounder(type)
   // Preferred: API request
